Key press handling functions now return a value
To allow one part of the UI to effect a change in another, event
handling functions now return an enum that desires the set of possible
changes. These changes are generally high level commands such as "switch
to the main screen" or "set the title of a task to X and then switch to
the main screen".

An alternative design would pass the entire application state to every
event handling function. This is the design the program used to have,
but this has a few disadvantages. Foremost, it is not a composable
design.

Another would be to pass these command messages by way of a stream or
"channel". It may come to this, but such a pattern would rely on side
effects. For now, I like the simplicity of a procedural style with
simple return values.

Thank you to the folks on the Ratatui forum for making these
considerations more clear to me:
https://forum.ratatui.rs/t/question-how-can-i-representing-applicatoin-state-ergonomically-in-ratatui-programs/54